chore(client): rename, rebrand, apply proper license attributions to package
