+++ internal/pinchergo/cache.go
@@ -1,4 +1,4 @@
-package client
+package pinchergo
 
 import (
 	"maps"
