+++ internal/pinchergo/endpoints.go
@@ -1,4 +1,4 @@
-package client
+package pinchergo
 
 import (
 	"fmt"
