+++ internal/pinchergo/requests_accounts.go
@@ -1,4 +1,4 @@
-package client
+package pinchergo
 
 import (
 	"net/http"
