+++ internal/pinchergo/requests_auth.go
@@ -1,4 +1,4 @@
-package client
+package pinchergo
 
 import (
 	"fmt"
