+++ internal/pinchergo/requests_categories.go
@@ -1,4 +1,4 @@
-package client
+package pinchergo
 
 import (
 	"net/http"
