+++ internal/pinchergo/requests_payees.go
@@ -1,4 +1,4 @@
-package client
+package pinchergo
 
 import (
 	"net/http"
