+++ internal/pinchergo/requests_state.go
@@ -1,4 +1,4 @@
-package client
+package pinchergo
 
 import (
 	"net/http"
