+++ internal/pinchergo/requests_txns.go
@@ -1,4 +1,4 @@
-package client
+package pinchergo
 
 import (
 	"net/http"
